server/requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
scikit-learn
llama-cpp-python
